Updating HighChartAppKernel Constructor & Call Sites (#50)
* Updating HighChartAppKernel Constructor & Call Sites

The HighChartAppKernel class needed to have an XDUser parameter added / piped to the parent constructor.

All instantiations of this class needed to have this new parameter added.

* Removing duplicate code / null style fix

Duplicate code was added during the rebase procedure @ line 472. This has been
removed.

There was also an instance of `NULL` which was changed to `null` per our style
guidelines.