Added CenterReportCardPortlet and required infrastructure
Added a new Portlet, the Center Report Card Portlet, for the Novice User view.
This portlet displays App Kernel run information ( Failed, Under Performing, In
Control, and Over Performing ) for a given time period. If the user clicks on
one of the presented rows, they will be taken to the `App Kernels-> Performance
Map` tab. It's duration toolbar will be updated to match the start / end date
the portlet is using and the row that corresponds to the resource / app kernel
the user clicked on will be selected.

- To support this new feature, the `Performance Map` was moved from the Internal
Dashboard to be a new sub tab of `App Kernels`.
- A new function, `getPerformanceMapRaw`, was added to
`AppKernelControllerProvider`.
- `AppKernels.js` was modified so that it knows how to activate a sub tab, if one
is provided.
- A number of masking / unmasking calls were added so that the user is kept in
the loop as to when XDMoD is busy retrieving data.