Novice User Interface - Center Report Card (#63)
* Added CenterReportCardPortlet and required infrastructure

Added a new Portlet, the Center Report Card Portlet, for the Novice User view.
This portlet displays App Kernel run information ( Failed, Under Performing, In
Control, and Over Performing ) for a given time period. If the user clicks on
one of the presented rows, they will be taken to the `App Kernels-> Performance
Map` tab. It's duration toolbar will be updated to match the start / end date
the portlet is using and the row that corresponds to the resource / app kernel
the user clicked on will be selected.

- To support this new feature, the `Performance Map` was moved from the Internal
Dashboard to be a new sub tab of `App Kernels`.
- A new function, `getPerformanceMapRaw`, was added to
`AppKernelControllerProvider`.
- `AppKernels.js` was modified so that it knows how to activate a sub tab, if one
is provided.
- A number of masking / unmasking calls were added so that the user is kept in
the loop as to when XDMoD is busy retrieving data.

* Updating build files

Just updating the build files so that they're included in the rpm.

* Updated to remove XSEDE dependency

The `Compliance` class is located in the `XSEDE` module but is utilized here in
the `AppKernels` module. I believe we want users to be able to utilize the
`AppKernels` module without having `XSEDE` installed, so this should help us
achieve that goal.

* Reformatted per eslint feedback

* Adding a CSS file to this modules assets

We were already including the `ColumnHeaderGroup.js` file here, now we also
include `ColumnHeaderGroup.css`.

* Added documentation to .travis file.

Just documenting the previously added line.

* Updates to allow additional filtering of app kernel data

Added the ability for the the AppKernel Performance Map to be filtered by a set
of authorized resources or the set of resources associated with a User's
organization.

* Removed userOrganization and cleaned up resource_ids

* Updates to the `getRawPerformanceMap` function

- Added some additional documentation / throws clause to reflect requiring the
  user calling this endpoint to have either the Center Director or Center Staff
  acl.
- Updated the function to require the authenticated user making the request to
  have the Center Director or Center Staff acl.
- Replaced the use of `MetricExplorer::getDimensionValues` with
  `XDUser::getResources` per research / discussion w/ @jpwhite4.

* Updating the authorization method for `getRawPerformanceMap`

We require that either Center Director's *or* Center Staff be allowed to utilize
this endpoint. The original `authorize($request, $requirements)` function call
instead ensures that the user have *all* of the acls supplied.

* Update authorization logic & resource parameter

- authorization logic update: derrrrr.... I r know logics.
- `resource` param for `PerformanceMap`: needed to massage the form it was
  provided to `PerformanceMap`.

* Fixing path to ColumnHeaderGroup.css

* Pinning dist so that php 5.4 will work

* eslint fix

* asfjlk

* Added documentation & formatted some sql

* Standardizing authorization logic

We have the same logic in `XDUser::getResources` just updating so that they're
the same.