Updates to allow for PO usage
This functionality was originally added for OpenXDMoD and as such did not take into
account that PO would also be utilizing this endpoint. Logic has been added to the
 options building that defaults to not restricting the resources
returned, but if the user does not have  then they are restricted to seeing only
the resources they have access to.