Attempting to resolve priv error during setup
Received the following error in CircleCI during setup `1Can not access
mod_appkernels mod_akrr databases.` which seems to indicate that the
`akrruser` does not have privs to the specified dbs. The user / db
creation script does seem to give the user privs to the dbs in question
so hopefully the `FLUSH PRIVILEGES` will resolve the problem.